requirements.txt
google-cloud-storage
apache-airflow
pytest
apache-beam[gcp]
scikit-learn
pandas